# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "教学资料文件：PDF / Word / PPT / 图片 / 视频 / 录音",
            "title": "File",
            "type": "string"
          },
          "is_reference": {
            "default": false,
            "description": "是否标记为参考资料：影响备课检索加权与生成物溯源",
            "title": "Is Reference",
            "type": "boolean"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post",
        "type": "object"
      },
...
      "DocumentChunk": {
        "description": "一个分块：`chunk_id` 全局唯一，`chunk_index` 是它在原文中的次序。",
        "properties": {
          "chunk_id": {
            "title": "Chunk Id",
            "type": "integer"
          },
          "chunk_index": {
            "title": "Chunk Index",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "chunk_index",
          "content"
        ],
        "title": "DocumentChunk",
        "type": "object"
      },
      "DocumentDetail": {
        "example": {
          "file_type": "pdf",
          "filename": "一次函数讲义.pdf",
          "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
          "is_reference": true,
          "status": "处理中"
        },
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunk"
            },
            "title": "Chunks",
            "type": "array"
          },
          "conflict_count": {
            "title": "Conflict Count",
            "type": "integer"
          },
          "failure_reason": {
            "default": "",
            "title": "Failure Reason",
            "type": "string"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_reference": {
            "title": "Is Reference",
            "type": "boolean"
          },
          "parsed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At"
          },
          "status": {
            "enum": [
              "处理中",
              "已完成",
              "有冲突",
              "失败"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "file_type",
          "status",
          "is_reference",
          "parsed_at",
          "conflict_count",
          "chunk_count",
          "chunks"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentListResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentView"
            },
            "title": "Documents",
            "type": "array"
          }
        },
        "required": [
          "documents"
        ],
        "title": "DocumentListResponse",
        "type": "object"
      },
      "DocumentView": {
        "description": "一份教学资料：上传响应与列表项共用的形状。",
        "example": {
          "file_type": "pdf",
          "filename": "一次函数讲义.pdf",
          "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
          "is_reference": true,
          "status": "处理中"
        },
        "properties": {
          "failure_reason": {
            "default": "",
            "title": "Failure Reason",
            "type": "string"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_reference": {
            "title": "Is Reference",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "处理中",
              "已完成",
              "有冲突",
              "失败"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "file_type",
          "status",
          "is_reference"
        ],
        "title": "DocumentView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReferenceRequest": {
        "description": "参考资料标记的目标值（幂等设置，不是翻转）。",
        "example": {
          "is_reference": true
        },
        "properties": {
          "is_reference": {
            "title": "Is Reference",
            "type": "boolean"
          }
        },
        "required": [
          "is_reference"
        ],
        "title": "ReferenceRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "多模态 AI 备课智能体后端。\n\nOpenAPI 是本服务接口文档的**唯一事实源**：每个端点都带 `summary`、描述、\n请求/响应示例、错误码与分组 tag；前端 TypeScript 类型由本 schema 生成，禁止手抄。\n* **界面**：Swagger UI `/docs`，ReDoc `/redoc`。\n* **分组**：系统 / 备课会话 / 知识库 / 知识图谱 / 生成物 / 冲突审核 / 题库。\n* **单用户**：固定 `user_id=\"default\"`，本服务不含登录与多租户。\n* **启动清扫**：进程重启时，上一轮被中断的「处理中」资料不会有后台任务再推进，\n  服务启动时统一标记为「失败」，避免前端无限轮询（教师可按原文重新上传）。\n* **能力切换**：对话模型、向量化、语音转写、PDF 解析、网络搜索均按配置选择。\n  云端能力未配置时相关端点返回 `503`（`code: provider_not_configured`，`message` 面向教师，\n  指到设置页）——产品不返回任何「看起来像真结果的占位内容」。\n* **生成失败**：配好了但生成没成功时返回 `502`（`code: generation_failed`），\n  教师可重试——不落一份空课件 / 空教案冒充成功。\n* **版本策略**：只有破坏性变更才升 `/api/v2`。\n",
    "title": "EduMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "description": "返回全部教学资料及其解析状态，供知识库工作台轮询进度。\n\n`status` 取值：`处理中` → 后台解析尚未结束；`已完成` → 已入库可检索；`有冲突` → 已入库但检测到待裁决的矛盾（见冲突审核区）；`失败` → 解析失败，可按原文重新上传。`is_reference` 为参考资料标记。\n\n含 `处理中` 的资料时，客户端应保持轮询直到全部落到终态；单份资料的完整信息（分块、冲突数、完成时间）见 `GET /api/v1/documents/{document_id}`。",
        "operationId": "list_documents_api_v1_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "documents": [
                    {
                      "file_type": "pdf",
                      "filename": "一次函数讲义.pdf",
                      "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
                      "is_reference": true,
                      "status": "已完成"
                    },
                    {
                      "file_type": "mp3",
                      "filename": "课堂录音.mp3",
                      "id": "1a2b3c4d-5e6f-4071-8293-a4b5c6d7e8f9",
                      "is_reference": false,
                      "status": "处理中"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "description": "资料列表"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "教学资料列表",
        "tags": [
          "知识库"
        ]
      }
    },
    "/api/v1/documents/upload": {
      "post": {
        "description": "上传一份教学资料并**立即返回**：文件落盘 + 建文档记录 + 交后台解析（解析完成前`status` 为 `处理中`）。\n\n上传后请轮询 `GET /api/v1/documents` 直到状态从 `处理中` 变为 `已完成`（或 `有冲突` / `失败`），不要靠等待本请求变慢来判断进度。\n\n`is_reference=true` 把文档标记为参考资料：发起备课时勾选其 id 会提高检索权重，并在生成物中溯源。后续可用 `PATCH /api/v1/documents/{document_id}/reference` 随时切换该标记。\n\n**入口即拒绝的两种情况**：扩展名不在可解析清单内返回 `415`（清单见错误说明）；空文件返回 `400`。这两种情况不落盘、不建记录——收下再让后台解析失败，教师只会看到一个没有原因的「失败」。\n\n解析失败时资料的 `failure_reason` 给出面向教师的一句原因（文件损坏 / 能力未配置 / 被中断）。",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "file_type": "pdf",
                  "filename": "一次函数讲义.pdf",
                  "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
                  "is_reference": true,
                  "status": "处理中"
                },
                "schema": {
                  "$ref": "#/components/schemas/DocumentView"
                }
              }
            },
            "description": "已接收并交后台解析（`status` 为 `处理中`）"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "文件是空的，请确认后重新选择"
                }
              }
            },
            "description": "空文件：没有可解析的内容"
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "文件过大"
                }
              }
            },
            "description": "文件过大：超过单次上传上限（默认 200 MB）"
          },
          "415": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "不支持的资料格式「exe」。可上传：pdf、doc、docx、ppt、pptx、png、jpg …"
                }
              }
            },
            "description": "不支持的资料格式：扩展名不在可解析清单内"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "上传教学资料",
        "tags": [
          "知识库"
        ]
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "description": "一份教学资料的全部信息：解析状态与参考资料标记（与列表同一口径），外加解析完成时间、待审冲突数，以及入库的分块明细（可回溯到来源资料）。\n\n`chunks` 只返回前 200 个分块，`chunk_count` 是该资料的分块总数；资料仍在 `处理中` 时 `parsed_at` 为空、`chunks` 为空——列表与详情的状态都会随后台解析自动变化，客户端无需手动刷新。",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "description": "教学资料 id，取自上传响应或资料列表",
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "description": "教学资料 id，取自上传响应或资料列表",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "chunk_count": 12,
                  "chunks": [
                    {
                      "chunk_id": 42,
                      "chunk_index": 0,
                      "content": "一次函数：形如 y=kx+b（k≠0）的函数。"
                    }
                  ],
                  "conflict_count": 0,
                  "file_type": "pdf",
                  "filename": "一次函数讲义.pdf",
                  "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
                  "is_reference": true,
                  "parsed_at": "2026-09-24T10:30:00",
                  "status": "已完成"
                },
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "教学资料详情"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "资料不存在: 7f6e5d4c-3b2a-4918-8776-655443322110"
                }
              }
            },
            "description": "资料不存在：id 不属于任何一份已上传的教学资料"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "教学资料详情",
        "tags": [
          "知识库"
        ]
      }
    },
    "/api/v1/documents/{document_id}/reference": {
      "patch": {
        "description": "把教学资料标记为参考资料（或取消标记），用于备课检索加权与生成物溯源。\n\n请求体给的是**目标值**而不是翻转动作，重复提交同一值结果一致（幂等）；响应即最新状态，切换无需等待任何后台任务。",
        "operationId": "set_reference_api_v1_documents__document_id__reference_patch",
        "parameters": [
          {
            "description": "教学资料 id，取自上传响应或资料列表",
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "description": "教学资料 id，取自上传响应或资料列表",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "file_type": "pdf",
                  "filename": "一次函数讲义.pdf",
                  "id": "7f6e5d4c-3b2a-4918-8776-655443322110",
                  "is_reference": true,
                  "status": "已完成"
                },
                "schema": {
                  "$ref": "#/components/schemas/DocumentView"
                }
              }
            },
            "description": "切换后的资料（标记已生效）"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "资料不存在: 7f6e5d4c-3b2a-4918-8776-655443322110"
                }
              }
            },
            "description": "资料不存在：id 不属于任何一份已上传的教学资料"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          }
        },
        "summary": "切换参考资料标记",
        "tags": [
          "知识库"
        ]
      }
    },
...
  },
...
}